requirements-dev.txt
PyYAML>=6.0
jsonschema>=4.20
pytest>=8.2
